+++ test/components/test_traffic_volume.py
@@ -10,11 +10,11 @@ from vcr import use_cassette
 
 from traffic_emissions.components.traffic_volume import (
     assign_traffic,
-    calculate_mean_pop_density_polygon,
     get_roads,
     get_scaling_factor,
     traffic_volume,
 )
+from traffic_emissions.components.utils import calculate_mean_pop_density_polygon
 
 LINE_GEOM = gpd.GeoSeries(
     [LineString([(325193.9834442865, 4669724.10105637), (325111.5841199331, 4669734.855883078)]) for _ in range(6)]
@@ -36,8 +36,7 @@ def test_get_scaling_factor(default_aoi):
 
 
 def test_calculate_mean_pop_density_polygon(default_aoi):
-    pop_path = 'resources/pop_dens.tif'
-    mean_pop_dens_aoi, pop_sum_aoi = calculate_mean_pop_density_polygon(pop_path, default_aoi)
+    mean_pop_dens_aoi, pop_sum_aoi = calculate_mean_pop_density_polygon(default_aoi)
     assert round(mean_pop_dens_aoi, 2) == 1358.02
     assert round(pop_sum_aoi, 0) == 74691